refactoring the code to use components with conditional rendering instead of routes
