+++ src/components/Logo.svelte
@@ -1,0 +1,7 @@
+<script>
+  import logo from "$assets/images/logo.png";
+</script>
+
+<div class="logo">
+  <img id="logo" width="65%" src={logo} alt="OpenGOAL" />
+</div>
