+++ src/routes/Jak1.svelte
@@ -1,39 +1,23 @@
 <script>
   // Assets
-  import logo from "$assets/images/logo.png";
-  import { onMount } from "svelte";
-  import { Link } from "svelte-routing";
   import { getInstallStatus, SupportedGame } from "$lib/config";
-  import { launchGame } from "$lib/launch";
+  import Setup from "../components/setup/Jak1Setup.svelte";
+  import Jak1Main from "../components/games/Jak1Main.svelte";
+  import Console from "../components/Console.svelte";
+  import Logo from "../components/Logo.svelte";
 
   // State
-  let gameInstalled = false;
-
-  // Events
-  onMount(async () => {
-    gameInstalled = await getInstallStatus(SupportedGame.Jak1);
-  });
-
-  function onClickConfig() {
-    alert("TODO");
-  }
-  function onClickPlay() {
-    launchGame();
-  }
+  const gameInstalled = async () => await getInstallStatus(SupportedGame.Jak1);
 </script>
 
 <div class="flex-center">
-  <div class="logo">
-    <img id="logo" width="65%" src={logo} alt="OpenGOAL" />
-  </div>
-  <div id="launcherControls">
-    {#if gameInstalled}
-      <button class="btn lg" on:click={onClickConfig} disabled>Config</button>
-      <button class="btn lg" on:click={onClickPlay}>Play</button>
+  <Logo />
+  {#await gameInstalled() then installed}
+    {#if installed}
+      <Jak1Main />
     {:else}
-      <Link to="/setup/jak1">
-        <button class="btn lg">Setup</button>
-      </Link>
+      <Setup />
+      <Console />
     {/if}
-  </div>
+  {/await}
 </div>
